requirements.txt
numpy>=1.24.1,<1.25.0
python-can>=4.1.0,<4.2.0
tomlkit>=0.11.6,<1.24.0
PyQt5>=5.15.7,<5.16.0
black>=22.12.0
pylint>=2.15.10,<5.16.0
